Add files via upload
Now-a-days communication plays a major role in everything be it professional or
personal. Email communication service is being used extensively because of its
free use services, low-cost operations, accessibility, and popularity. Emails have
one major security flaw that is anyone can send an email to anyone just by getting
their unique user id. 
A tool
that can detect and classify spam. In addition to that, it also provides information
regarding the text provided in a quick view format for user convenience.